+++ src/cprotocol.cpp
@@ -172,7 +172,7 @@ CPacketStream *CProtocol::GetStream(uint16 uiStreamId, const CIp *Ip)
     CPacketStream *stream = NULL;
     
     // find if we have a stream with same streamid in our cache
-    for ( int i = 0; (i < m_Streams.size()) && (stream == NULL); i++ )
+    for ( size_t i = 0; (i < m_Streams.size()) && (stream == NULL); i++ )
     {
         if ( m_Streams[i]->GetStreamId() == uiStreamId )
         {
@@ -189,7 +189,7 @@ CPacketStream *CProtocol::GetStream(uint16 uiStreamId, const CIp *Ip)
 
 void CProtocol::CheckStreamsTimeout(void)
 {
-    for ( int i = 0; i < m_Streams.size(); i++ )
+    for ( size_t i = 0; i < m_Streams.size(); i++ )
     {
         // time out ?
         m_Streams[i]->Lock();
